+++ yyt.config.js
@@ -1,0 +1,4 @@
+module.exports = {
+  src_folders: ['nightwatch/test'],
+  output_folder: false
+};